Relocate exceptions, update "since" tag and license
