+++ cas-server-support-osf/src/main/java/io/cos/cas/adaptors/postgres/handlers/OpenScienceFrameworkAuthenticationHandler.java
@@ -24,11 +24,11 @@ import io.cos.cas.adaptors.postgres.models.OpenScienceFrameworkTimeBasedOneTimeP
 import io.cos.cas.adaptors.postgres.models.OpenScienceFrameworkUser;
 import io.cos.cas.authentication.exceptions.AccountNotConfirmedIdPLoginException;
 import io.cos.cas.authentication.exceptions.AccountNotConfirmedOsfLoginException;
-import io.cos.cas.authentication.InvalidVerificationKeyException;
-import io.cos.cas.authentication.OneTimePasswordFailedLoginException;
-import io.cos.cas.authentication.OneTimePasswordRequiredException;
+import io.cos.cas.authentication.exceptions.InvalidVerificationKeyException;
+import io.cos.cas.authentication.exceptions.OneTimePasswordFailedLoginException;
+import io.cos.cas.authentication.exceptions.OneTimePasswordRequiredException;
 import io.cos.cas.authentication.OpenScienceFrameworkCredential;
-import io.cos.cas.authentication.ShouldNotHappenException;
+import io.cos.cas.authentication.exceptions.ShouldNotHappenException;
 import io.cos.cas.authentication.oath.TotpUtils;
 
 import org.jasig.cas.authentication.AccountDisabledException;
