+++ cas-server-support-osf/src/main/java/io/cos/cas/authentication/exceptions/RemoteUserFailedLoginException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016. Center for Open Science
+ * Copyright (c) 2020. Center for Open Science
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cos.cas.authentication;
+package io.cos.cas.authentication.exceptions;
 
 import javax.security.auth.login.AccountException;
 
@@ -22,7 +22,7 @@ import javax.security.auth.login.AccountException;
  *
  * @author Michael Haselton
  * @author Longze Chen
- * @since 4.1.5
+ * @since 20.1.0
  */
 public class RemoteUserFailedLoginException extends AccountException {
