+++ cas-server-support-osf/src/test/java/io/cos/cas/authentication/handler/support/OpenScienceFrameworkPrincipalFromRequestRemoteUserNonInteractiveCredentialsActionTests.java
@@ -3,7 +3,7 @@ package io.cos.cas.authentication.handler.support;
 import io.cos.cas.AbstractTestUtils;
 import io.cos.cas.adaptors.postgres.types.DelegationProtocol;
 import io.cos.cas.authentication.OpenScienceFrameworkCredential;
-import io.cos.cas.authentication.RemoteUserFailedLoginException;
+import io.cos.cas.authentication.exceptions.RemoteUserFailedLoginException;
 import io.cos.cas.mock.MockNormalizeRemotePrincipal;
 import io.cos.cas.mock.MockNotifyRemotePrincipalAuthenticated;
